Built the Candidate List page
Created a candidates controller that sends a static address to the google factory, and passes it to the candidates scope. Candidates html file takes that scope and generates a list of names, they are supposed to link to specific pages for that individual candidate, but that seems to be broken.