Created Polling Location portion of the app.
Created the Controller and reference in the app.js file, and setup a basic template for it. However I can't seem to test it properly, the address info isn't passing to the template correctly. When I test the URL in the browser, I'm getting back errors anyways. API might be temporarily down.